pyproject.toml
[tool.poetry]
name = "chatgpt-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Umbreella <drjagin-vadim@rambler.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-telegram-bot = { version = "20.1", extras = ["rate-limiter"] }
openai = "0.28.1"
tiktoken = ">=0.3.0"
pymongo = "4.3.3"
python-dotenv = "0.21.0"
yookassa = "3.4.2"
pydantic-settings = "^2.7.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
exclude = [
    './bot/**',
]

[tool.ruff.lint]
ignore = ["F401"]

[tool.ruff.format]
quote-style = "single"